refactor(server): group the client-facing response normalization
The response the server writes back is normalized against the query as the
client sent it, not against request.Req, which the resolver chain mutates in
place. That was two captured locals and three statements spread across
resolve(); collect them in a clientQuery value captured before resolution and
applied afterwards, so further per-client normalization has one home.

No behavior change: the existing server specs pass unchanged.
